% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handsometable.R
\name{handsome_table}
\alias{handsome_table}
\title{Handsome Table}
\usage{
handsome_table(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{data.frame to pass into the widget}
}
\description{
An R Handsontable Wrapper
}
